relocate original files
